Add command services

+++ src/main/java/com/hsmnzaydn/termcommands/Command/CommandRepository.java
@@ -2,5 +2,5 @@ package com.hsmnzaydn.termcommands.Command;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommandRepository extends JpaRepository<Command,Long> {
+public interface CommandRepository extends JpaRepository<Command,Integer> {
 }
